GhostAI

Documentation

listLength :: [a] -> Int

Length of a list. Used for the score

checkInList :: Eq a => a -> [a] -> Bool

Reverse the True and False statements from the standard elem function

ghostModeTimer :: PacGame -> PacGame

Changing from scatter to chase follow Pac-Mans rules.

ghostModeSwitch :: PacGame -> PacGame

Determines whether or not pacman ate the power pellet.

releaseInky :: PacGame -> PacGame

Function to release Inky. Originally tried to fit them all in one, but it did not work

releaseClyde :: PacGame -> PacGame

Function to release Clyde.

releasePinky :: PacGame -> PacGame

Function to release Pinky, Inky, and Clyde at their appropriate times.

distance :: Point -> Point -> Float

Calculate distance between 2 points

ghostCollision :: PacGame -> PacGame

Checks if you ran into any of the ghosts. Decreases a life it above 0. Sets gameStatus to LOST if out of lives.

ghostCollisionSwitch :: PacGame -> PacGame

Checks if you ran into any of the ghosts. Decreases a life it above 0. Sets gameStatus to LOST if out of lives.

checkEaten :: PacGame -> Int

moveToTargetB :: PacGame -> PacGame

Moves Blinky to the target...which is either the top right corner or pacman

moveToTargetP :: PacGame -> PacGame

Moves Pinky to the target...which is either the top left corner or a few spaces in front of pacman

targetInky :: PacGame -> Point

Determines Inky's target based on the location && direction of Pacman and the location of Blinky

moveToTargetI :: PacGame -> PacGame

Moves Inky to the target

targetClyde :: PacGame -> Point

Computes the proper target of clyde. If Clyde gets within a 4 block radius, sets target to clydescattertarget

moveToTargetC :: PacGame -> PacGame

Moves Clyde to the target. Doesn't come out until at least 50 pellets are eaten! (roughly 1/3 total)

findWalls :: PacGame -> Int -> Int

Used for chase AI, Checks current position with walls, returning a number which represents how and where the walls are.