Main

Documentation

handleKeys :: Event -> PacGame -> PacGame

Handle user inputs. Needed for gloss play function

initialize :: PacGame

Initialize base game state

beginTimer :: PacGame -> PacGame

Timer that runs to let me know when I can release the other ghosts into the maze.

checkCollision :: Point -> Bool

Checks if point given is within the points that create the walls of the maze

renderPacman :: PacGame -> Picture

Renders pacman's image based on the gamestate.

renderPinky :: PacGame -> Picture

Renders pinky's image based on the gamestate.

renderInky :: PacGame -> Picture

Renders Inky's image based on the gamestate.

renderBlinky :: PacGame -> Picture

Renders Blinky's image based on the gamestate.

renderClyde :: PacGame -> Picture

Renders Clyde's image based on the gamestate.

renderStuff :: PacGame -> Picture

renderPellets :: PacGame -> Picture

Renders the pellets images based on the gamestate.

renderPowerPellets :: PacGame -> Picture

Renders the power pellet images based on the gamestate.

renderLives :: PacGame -> Picture

Render the lives on the bottom of the screen

displayWindow :: Display

Main display window.

renderVictory :: Picture

Renders the victory screen

renderDefeat :: Picture

Renders the defeat screen

render :: PacGame -> Picture

Main render function used in play. List of multiple other render functions that layer on one another.

main :: IO ()

Main play function of the game

sMove :: Float -> PacGame -> PacGame

This is the standard movement function. Take the velocity and position, then update it with (position + velocity * time)

notMoving :: PacGame -> Bool

Returns true if pacman is stopped.

execM :: PacGame -> PacGame

Function that handles pacmans movements

wallCollision :: PacGame -> PacGame

Collision checks on walls

pacEat :: PacGame -> PacGame

Function that checks if pacman is over a pellet. If so then remove it from the list held in the gamestate

pacEatP :: PacGame -> PacGame

Function that checks if pacman is over a power pellet. If so then remove it from the list held in the gamestate and change mode to frightened

pacScore :: PacGame -> PacGame

Determines the current score of your game. Regular pellets are worth 10 points. Large Pellets are worth 50

checkTeleport :: PacGame -> PacGame

Check if pacman is at teleport location. If so, teleport him to the other side

ghostCollision :: PacGame -> PacGame

Checks if you ran into any of the ghosts. Decreases a life it above 0. Sets gameStatus to LOST if out of lives.