Maze

Documentation

coords :: [Point]

Defines the list of points that make up the maze of Pacman

pelletsL :: [Point]

Defines the list of pellets that are around the maze.

powerPellet :: [Point]

List of the 4 power pellets

crossR :: [Point]

List of all the corssroads for the ghosts to make movement choices.

renderS :: Point -> Picture

Turns a single point into a block picture to be displayed on the map

renderW :: [Point] -> Picture

Takes the list of points turns them into the Pacman maze

renderP :: Point -> Picture

Turns a single point into a pellet picture to be displayed on the map

renderPS :: [Point] -> Picture

Takes the list of points turns them into the pellets

renderPP :: Point -> Picture

Turns a single point into a power pellet picture to be displayed on the map

renderPPS :: [Point] -> Picture

Takes the list of points turns them into the power pellets