Structs

Documentation

type Velocity = (Float, Float)

Velocity of the ghost's and pacman. Used for movement

fps :: Int

FPS of the game. Default is 60

frightenLength :: Float

How long the ghosts will go into frighten mode

data Name

Data describing the different names of the PacMan Characters

Constructors

Pacman

Pinky

Inky

Blinky

Clyde

data Direction

Data describing the directions of Pacman

Constructors

UP

DOWN

LEFT

RIGHT

STOP

data GameStatus

Data describing win/loss

Constructors

WON

LOST

PLAYING

data Mode

Data for ghost modes (Scatter, Chase, Frightened)

Constructors

SCATTER

CHASE

FRIGHTENED

blinkyScatterTarget :: Point

Blinky's corner of the map

inkyScatterTarget :: Point

Inky's corner of the map

pinkyScatterTarget :: Point

Pinky's corner of the map

clydeScatterTarget :: Point

Clyde's corner of the map

data Characters

Data describing Pacman

Constructors

Characters

Fields

cName :: Name

Name of character from Name types

speed :: Velocity

Constant speed of character

location :: Point

Location of character in Point form

data Ghost

Data describing the Ghosts (Blinky, Pinky, Inky, Clyde)

Constructors

Ghost

Fields

gName :: Name

gSpeed :: Velocity

gLocation :: Point

gTarget :: Point

gLastMove :: Direction

gDirection :: Direction

data PacGame

Data describing the state of the game

Constructors

Game

Fields

lives :: Int

Number of lives the Pacman has

gameStatus :: GameStatus

Status of the current game (WON/LOSS)

pacman :: Characters

Pacman character data

pinky :: Ghost

Pinky character data

inky :: Ghost

blinky :: Ghost

clyde :: Ghost

time :: Float

Used to determine when to release the other 3 characters

score :: Int

Keeps track of how many pellets have been eaten!

pellets :: [Point]

List of pellet locations in the maze

direction :: Direction

Current direction of Pacman

bufDirection :: Direction

Buffered direction of movement

gMode :: Mode

Mode of the ghosts, which can either be Scatter, Chase, Frightened

lcrB :: Point

Last cross road for Blinky

lcrP :: Point

Last cross road for Pinky

lcrI :: Point

Last cross road for Inky

lcrC :: Point

Last cross road for Clyde

pPellets :: [Point]

List of the 4 power pellet locations in the maze

fTime :: Float

Time that ghost go into frightened mode for (5 seconds)